+++ include/mruby_integration/models/window.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "mruby.h"
+
+void
+append_models_Window(mrb_state*);