+++ mrb_doc/core/timing.rb
@@ -23,7 +23,7 @@ def get_frame_time
   0.01633
 end
 
-# Returns the amount of time the has passed since {init_window} was called.
+# Returns the amount of time the has passed since {Window.open} was called.
 # @return [Float]
 def get_time
   # mrb_time