+++ src/ruby/models/window.rb
@@ -1,0 +1,4 @@
+# This class is used for managing your {Window} and retrieving any information
+# about it.
+class Window
+end